% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{c_bubble}
\alias{c_bubble}
\title{Bubble}
\usage{
c_bubble(c, serie, size = NULL, ..., label = NULL)
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}

\item{serie}{The bare name of the serie to plot on the y axis.}

\item{size}{The bare name of the variable defining the size of the points.}

\item{...}{Any other options.}

\item{label}{Label of the serie, if \code{NULL} then the \code{serie} is used.}
}
\description{
Add a bubble chart.
}
\examples{
mtcars \%>\% 
 c_hart(mpg) \%>\% 
 c_bubble(qsec, wt)

}
