% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{c_events}
\alias{c_events}
\title{Events}
\usage{
c_events(c)
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}
}
\description{
Capture how users interact with the chart.
}
