% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{c_layout}
\alias{c_layout}
\title{Layout}
\usage{
c_layout(c, left = NULL, right = NULL, top = NULL, bottom = NULL)
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}

\item{left, right, top, bottom}{Values, in pixels, defining margin around chart.}
}
\description{
Adjust the layout of the graph.
}
\examples{
cars \%>\% 
 c_hart(speed) \%>\% 
 c_scatter(dist) \%>\% 
 c_layout(
   top = 20, right = 20, 
   left = 20, bottom = 20
 )

}
