% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{c_legend}
\alias{c_legend}
\title{Legend}
\usage{
c_legend(
  c,
  display = TRUE,
  position = c("top", "left", "bottom", "right"),
  align = c("center", "start", "end"),
  full_width = TRUE,
  reverse = FALSE,
  rtl = FALSE
)
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}

\item{display}{Whether to show the legend.}

\item{position}{The position of the legend.}

\item{align}{Alignment of the legend.}

\item{full_width}{Marks that this box should take the
full width of the canvas (pushing down other boxes).}

\item{reverse}{Legend will show datasets in reverse order.}

\item{rtl}{Renders legend from right to left.}
}
\description{
Customise the legend.
}
\examples{
c_hart(cars, caes(speed, dist)) \%>\% 
 c_line() \%>\% 
 c_legend(position = "left")

}
