% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{c_legend_label}
\alias{c_legend_label}
\title{Legend Label}
\usage{
c_legend_label(
  c,
  box_width = 40,
  font_size = 12,
  font_style = "normal",
  font_family = helvetica(),
  padding = 10,
  use_point_style = FALSE
)
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}

\item{box_width}{Width of colored box in pixels.}

\item{font_size}{Font size of text in pixels.}

\item{font_style}{Font style.}

\item{font_family}{Font family to use.}

\item{padding}{Padding, in pixels, between rows of boxes.}

\item{use_point_style}{Label style will match corresponding point
style (size is based on the mimimum value between boxWidth and fontSize).}
}
\description{
Customise the legend labels.
}
\examples{
cars \%>\% 
 c_hart(speed) \%>\% 
 c_line(dist) \%>\% 
 c_legend(position = "bottom") \%>\% 
 c_legend_label(box_width = 100)

}
