% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{c_line}
\alias{c_line}
\title{Line}
\usage{
c_line(c, ..., label = NULL, inherit_caes = TRUE, data = NULL)
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}

\item{...}{Any other options.}

\item{label}{Label of the serie, if \code{NULL} then the \code{serie} is used.}

\item{inherit_caes}{Whether to inherit \code{\link{caes}} from \code{\link{c_hart}}.}

\item{data}{If \code{NULL} inherits data from \code{\link{c_hart}}.}
}
\description{
Add a line chart.
}
\examples{
c_hart(cars, caes(speed, dist)) \%>\% 
 c_line()

}
