% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{c_pie}
\alias{c_pie}
\title{Pie}
\usage{
c_pie(c, serie, ..., label = NULL)
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}

\item{serie}{The bare serie to plot on the y axis.}

\item{...}{Any other options.}

\item{label}{Label of the serie, if \code{NULL} then the \code{serie} is used.}
}
\description{
Add a pie chart.
}
\examples{
df <- data.frame(
 x = letters[1:5],
 y = runif(5, 1, 5)
)

df \%>\% 
 c_hart(x) \%>\% 
 c_pie(y)

}
