% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{c_radar}
\alias{c_radar}
\title{Radar}
\usage{
c_radar(c, serie, ..., label = NULL)
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}

\item{serie}{The bare name of the serie to plot on the y axis.}

\item{...}{Any other options.}

\item{label}{Label of the serie, if \code{NULL} then the \code{serie} is used.}
}
\description{
Add a radar chart.
}
\examples{
df <- data.frame(
 x = letters,
 y = runif(26)
)

df \%>\% 
 c_hart(x) \%>\% 
 c_radar(y)

}
