% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{c_title}
\alias{c_title}
\title{Title}
\usage{
c_title(
  c,
  title = "",
  position = c("top", "left", "bottom", "right"),
  font_size = 12,
  font_family = helvetica(),
  font_color = "#666",
  font_style = "bold",
  padding = 10,
  line_height = 1.2,
  display = TRUE
)
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}

\item{title}{Title to display, if a \code{list} or a \code{vector} of
length > 1 text is rendered on multiple lines.}

\item{position}{The position of the title.}

\item{font_size}{Font size of text in pixels.}

\item{font_family}{Font family to use.}

\item{font_color}{Color of text.}

\item{font_style}{Font style.}

\item{padding}{Padding, in pixels, between rows of boxes.}

\item{line_height}{Height of an individual line of text.}

\item{display}{Whether to display the title.}
}
\description{
Add a title to the chart.
}
\examples{
cars \%>\% 
 c_hart(speed) \%>\% 
 c_line(dist) \%>\% 
 c_title(
   c("Title", "Subtitle")
 )

}
