% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{c_tooltip}
\alias{c_tooltip}
\title{Tooltip}
\usage{
c_tooltip(
  c,
  enabled = TRUE,
  mode = c("nearest", "point", "index", "dataset", "x", "y"),
  intersect = TRUE,
  position = c("average", "nearest"),
  bg_color = "rgba(0, 0, 0, 0.8)",
  title_font_family = helvetica(),
  title_font_size = 12,
  title_font_style = "bold",
  title_font_color = "#fff",
  title_align = c("left", "right", "center"),
  title_spacing = 2,
  title_margin_bottom = 6,
  body_font_family = helvetica(),
  body_font_size = 12,
  body_font_style = "normal",
  body_font_color = "#fff",
  body_align = title_align,
  body_spacing = 2,
  footer_font_family = helvetica(),
  footer_font_size = 12,
  footer_font_style = "bold",
  footer_font_color = "#fff",
  footer_align = title_align,
  footer_spacing = 2,
  footer_margin_top = 6,
  x_padding = 6,
  y_padding = x_padding,
  caret_padding = 2,
  caret_size = 5,
  corner_radius = 6,
  multi_key_bg = "#fff",
  display_colors = TRUE,
  border_color = "rgba(0, 0, 0, 0)",
  border_width = 0,
  rtl = FALSE
)
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}

\item{enabled}{Whether to enable tooltips.}

\item{mode}{Defines which elements appear on the tooltip.}

\item{intersect}{If code{TRUE}, the tooltip mode applies only when the
mouse position intersects with an element. If \code{FALSE}, the mode will
be applied at all times.}

\item{position}{Where to position the tooltip.}

\item{bg_color}{Background color of the tooltip.}

\item{title_font_family}{Font of title.}

\item{title_font_size}{Font size of title.}

\item{title_font_style}{Font style of title.}

\item{title_font_color}{Font color of title.}

\item{title_align}{Horizontal alignment of the title text lines.}

\item{title_spacing}{Spacing to add to top and bottom of each title line.}

\item{title_margin_bottom}{Margin to add on bottom of title section.}

\item{body_font_family}{Body line font.}

\item{body_font_size}{Body font size.}

\item{body_font_style}{Body font style.}

\item{body_font_color}{Body font color.}

\item{body_align}{Horizontal alignment of the body text lines.}

\item{body_spacing}{Spacing to add to top and bottom of each tooltip item.}

\item{footer_font_family}{Footer font.}

\item{footer_font_size}{Footer font size.}

\item{footer_font_style}{Footer font style.}

\item{footer_font_color}{Footer font color.}

\item{footer_align}{Horizontal alignment of the footer text lines.}

\item{footer_spacing}{Spacing to add to top and bottom of each footer line.}

\item{footer_margin_top}{Margin to add before drawing the footer.}

\item{x_padding, y_padding}{Padding to add on left and right of tooltip.}

\item{caret_padding}{Extra distance to move the end of the tooltip arrow
away from the tooltip point.}

\item{caret_size}{Size, in px, of the tooltip arrow.}

\item{corner_radius}{Radius of tooltip corner curves.}

\item{multi_key_bg}{Color to draw behind the colored boxes when multiple
items are in the tooltip.}

\item{display_colors}{If \code{TRUE}, color boxes are shown in the tooltip.}

\item{border_color}{Color of the border.}

\item{border_width}{Size of the border.}

\item{rtl}{Renders the tooltip left to right.}
}
\description{
Customise the tooltip.
}
\examples{
cars \%>\% 
 c_hart(speed) \%>\% 
 c_scatter(dist) \%>\% 
 c_tooltip(FALSE)

}
