% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caes.R
\name{caes}
\alias{caes}
\title{Aesthetics}
\usage{
caes(x, y, ...)
}
\arguments{
\item{x, y, ...}{List of name value pairs giving aesthetics to map to
variables. The names for x and y aesthetics are typically omitted because
they are so common; all other aspects must be named.}
}
\description{
Aesthetics to use to draw series on chart.
}
