% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{charter-shiny}
\alias{charter-shiny}
\alias{charterOutput}
\alias{renderCharter}
\alias{render_charter}
\alias{charterProxy}
\alias{charter_proxy}
\title{Shiny bindings for charter}
\usage{
charterOutput(outputId, width = "100\%", height = "400px")

renderCharter(expr, env = parent.frame(), quoted = FALSE)

render_charter(expr, env = parent.frame(), quoted = FALSE)

charterProxy(id, session = shiny::getDefaultReactiveDomain())

charter_proxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a charter}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using charter within Shiny
applications and interactive Rmd documents.
}
