% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_bar_caes}
\alias{error_bar_caes}
\title{Get Aesthetics from Type}
\usage{
error_bar_caes(type)
}
\arguments{
\item{type}{An error bar type.}
}
\description{
Get valid aesthetics from error bar type.
}
\keyword{internal}
