% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_bar_type}
\alias{error_bar_type}
\title{Convert Error Type}
\usage{
error_bar_type(type)
}
\arguments{
\item{type}{An error bar type.}
}
\description{
Convert convenient user facing type to internal error type.
}
\keyword{internal}
