% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{events}
\alias{events}
\alias{c_events}
\alias{c_events_capture}
\title{Events}
\usage{
c_events(
  c,
  events = c("mousemove", "mouseout", "click", "touchstart", "touchmove")
)

c_events_capture(c, events = c("click", "hover"))
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}

\item{events}{Events to register in the tooltip (\code{c_events})
or events to capture in Shiny server \code{c_events_capture}.}
}
\description{
Capture how users interact with the chart.
}
\examples{
library(shiny)
ui <- fluidPage(
  charterOutput("chart"),
 verbatimTextOutput("ev")
)

server <- function(input, output){

 output$chart <- render_charter({
   chart(cars, caes(speed, dist)) \%>\% 
     c_scatter() \%>\% 
     c_events_capture("hover")
 })

 output$ev <- renderPrint({
   input$chart_hovered
 })
}

if(interactive())
 shinyApp(ui, server)

}
