% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_serie}
\alias{generate_serie}
\title{Generate serie}
\usage{
generate_serie(
  c,
  data,
  label,
  inherit_caes,
  type = "line",
  ...,
  valid_caes = ALL_CAES
)
}
\arguments{
\item{c}{charter object.}

\item{data}{A data.frame.}

\item{label}{Serie label defined by user.}

\item{inherit_caes}{Whether to inherit aesthetics.}

\item{type}{Series type.}

\item{...}{Additional aesthetics and options.}
}
\description{
Wrapper to \code{handle_labels} and \code{make_serie}.
}
\keyword{internal}
