% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_label}
\alias{get_label}
\title{Get Label}
\usage{
get_label(data, label, caes, N)
}
\arguments{
\item{data}{Data.frame.}

\item{label}{Label set by user.}

\item{caes}{Aesthetics.}

\item{N}{Number of series (by group).}
}
\description{
Get label from of serie from data.
}
\keyword{internal}
