% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_to_serie}
\alias{group_to_serie}
\title{Groups to Series}
\usage{
group_to_serie(
  group_data,
  label,
  opts,
  N,
  type,
  caes,
  valid_caes = ALL_CAES,
  x_as_list = FALSE
)
}
\arguments{
\item{group_data}{A data.frame.}

\item{label}{Label assigned by user.}

\item{opts}{Additional options.}

\item{N}{Number of series.}

\item{type}{type of chart to draw.}

\item{caes}{Aesthetics.}

\item{valid_caes}{Valid aesthetics to keep in dataset.}

\item{x_as_list}{Whether to plot each x as a sublist.}
}
\description{
Maps groups to series.
}
\keyword{internal}
