% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_scales}
\alias{handle_scales}
\alias{handle_scale}
\title{Handle Scales}
\usage{
handle_scale(
  data = NULL,
  caes = list(),
  scales = list(),
  ...,
  axis = c("x", "y")
)

handle_scales(
  data = NULL,
  caes = list(),
  scales = list(),
  ...,
  axis = c("x", "y")
)
}
\arguments{
\item{data}{A data.frame.}

\item{caes}{Aesthetics.}

\item{scales}{Currently set scales.}

\item{...}{any other options to pass to axis.}

\item{axis}{Axis to define.}
}
\description{
Constructs axis based on input data.
}
\keyword{internal}
