% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listize}
\alias{listize}
\title{Listize}
\usage{
listize(data, valid_caes, x_as_list = FALSE)
}
\arguments{
\item{data}{A data.frame.}

\item{valid_caes}{Valid aesthetics.}

\item{x_as_list}{Whether to plot each x as a sublist.}
}
\description{
Turn a data.frame into a list
}
\section{X as list}{

Default data format is
\link{{x:1, y:2}, {x:1, y:3}, {x:2, y:3}, {x:2, y:5}}
with \code{x_as_list} set to \code{TRUE} each x is as a list
so the data looks like:
[\link{2,3},\link{3,5}]
Currently used for boxplot and violin plot.
}

\keyword{internal}
