% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_serie}
\alias{make_serie}
\title{Make Serie}
\usage{
make_serie(
  main_caes,
  main_data,
  data = NULL,
  inherit_caes = TRUE,
  type = "line",
  label = NULL,
  ...
)
}
\arguments{
\item{main_caes}{Main aesthetics.}

\item{main_data, data}{Data.frames.}

\item{inherit_caes}{Whether aethetics are inherited.}

\item{type}{Type of serie to make.}

\item{label}{Serie label.}

\item{...}{Addiitonal options and aesthetics.}
}
\description{
Make Chart.js series from aesthetics and data.
}
\keyword{internal}
