% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter.R
\name{c_hart}
\alias{c_hart}
\alias{chart}
\title{Initialise}
\usage{
c_hart(data = NULL, ..., width = NULL, height = NULL, elementId = NULL)

chart(data = NULL, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{Data.frame holding data to plot.}

\item{...}{Chart aesthetics, see \code{\link{caes}}.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Valid CSS id of chart to create.}
}
\description{
Initialise a chart.js chart
}
