% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{c_line}
\alias{c_line}
\title{Line}
\usage{
c_line(c, serie, ..., label = NULL)
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}

\item{serie}{The bare name of the serie to plot on the y axis.}

\item{...}{Any other options.}

\item{label}{Label of the serie, if \code{NULL} then the \code{serie} is used.}
}
\description{
Add a line chart.
}
\examples{
cars \%>\% 
 c_hart(dist) \%>\% 
 c_line(speed)

}
