% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_labels}
\alias{handle_labels}
\title{Handle Labels}
\usage{
handle_labels(
  labels = NULL,
  main_caes = list(),
  main_data = NULL,
  data = NULL,
  inherit_caes = TRUE,
  ...
)
}
\arguments{
\item{labels}{Current labels.}

\item{main_caes}{Aesthetics.}

\item{main_data, data}{Data.frames.}

\item{inherit_caes}{Whether to inherit aesthetics.}

\item{...}{Additional aesthetics.}
}
\description{
Handle global labels.
}
\keyword{internal}
