% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listize}
\alias{listize}
\title{Listize}
\usage{
listize(data, valid_caes)
}
\arguments{
\item{data}{A data.frame.}

\item{valid_caes}{Valid aesthetics.}
}
\description{
Turn a data.frame into a list
}
\keyword{internal}
