% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{series}
\alias{series}
\alias{c_line}
\alias{c_scatter}
\alias{c_bubble}
\alias{c_bar}
\alias{c_radar}
\alias{c_pie}
\alias{c_doughnut}
\alias{c_polar_area}
\alias{c_boxplot}
\alias{c_violin}
\alias{c_error_bar}
\title{Serie}
\usage{
c_line(c, ..., label = NULL, inherit_caes = TRUE, data = NULL)

c_scatter(c, ..., label = NULL, inherit_caes = TRUE, data = NULL)

c_bubble(c, ..., label = NULL, inherit_caes = TRUE, data = NULL)

c_bar(
  c,
  ...,
  label = NULL,
  inherit_caes = TRUE,
  data = NULL,
  orient = c("vertical", "horizontal")
)

c_radar(c, ..., label = NULL, inherit_caes = TRUE, data = NULL)

c_pie(c, ..., label = NULL, inherit_caes = TRUE, data = NULL)

c_doughnut(c, ..., label = NULL, inherit_caes = TRUE, data = NULL)

c_polar_area(c, ..., label = NULL, inherit_caes = TRUE, data = NULL)

c_boxplot(
  c,
  ...,
  label = NULL,
  inherit_caes = TRUE,
  data = NULL,
  orient = c("vertical", "horizontal")
)

c_violin(
  c,
  ...,
  label = NULL,
  inherit_caes = TRUE,
  data = NULL,
  orient = c("vertical", "horizontal")
)

c_error_bar(
  c,
  ...,
  label = NULL,
  inherit_caes = TRUE,
  data = NULL,
  type = NULL
)
}
\arguments{
\item{c}{An object of class \code{charter} as returned by \code{\link{c_hart}}.}

\item{...}{Any other options.}

\item{label}{Label of the serie, if \code{NULL} then the \code{serie} is used.}

\item{inherit_caes}{Whether to inherit \code{\link{caes}} from \code{\link{c_hart}}.}

\item{data}{If \code{NULL} inherits data from \code{\link{c_hart}}.}

\item{orient}{Orientation of box, violin, and plots.}

\item{type}{Type of error bar to apply.}
}
\description{
Add a series to chart.
}
\examples{
chart(cars, caes(speed, dist)) \%>\% 
 c_line()

mtcars \%>\% 
 chart(caes(qsec, mpg, group = cyl)) \%>\% 
 c_scatter(color = "red")

mtcars \%>\% 
 chart(caes(qsec, mpg, group = cyl)) \%>\% 
 c_bubble(caes(size = drat)) \%>\% 
 c_color_scheme("brewer.Oranges3")

chart(mtcars, caes(y = qsec, group = cyl)) \%>\% 
 c_violin() \%>\% 
 c_colors()

df <- data.frame(
 x = factor(c(1, 2)),
 y = c(1, 5),
 upper = c(1.1, 5.3),
 lower = c(0.8, 4.3)
)

chart(df, caes(x, y, xmin = lower, ymax = upper)) \%>\% 
 c_error_bar(type = "bar")

}
